refactor: split Transaction.Executor into EEVM.Handler.* phases
Phase 2 of the revm-style layering refactor.

The 562-line Transaction.Executor "god module" is replaced by a thin
orchestrator (EEVM.Handler) that delegates each pipeline phase to a
sibling sub-module — the same shape revm uses in its handler/ crate:

  EEVM.Handler.Validation     decode -> recover_sender -> validate
  EEVM.Handler.PreExecution   charge_upfront -> intrinsic + nonce bump
                              + access-list pre-warming + fee math
  EEVM.Handler.Execution      run top-level CALL or CREATE through the EVM
  EEVM.Handler.PostExecution  settle refund + coinbase reward + build result

The Handler.execute/4 entry point reads as a clean `with` chain that walks
the phases in order. The old EEVM.Transaction.Executor module is removed;
its corresponding test file is renamed to test/handler_test.exs.

No behaviour changes — 602 tests pass.
