+++ lib/eevm/interpreter.ex
@@ -3,7 +3,7 @@ defmodule EEVM.Interpreter do
   The EVM interpreter — fetches, decodes, and dispatches opcodes in a loop.
 
   This is the core opcode-execution engine. It is invoked by the
-  `EEVM.Transaction.Executor` (the transaction-level pipeline) and by system
+  `EEVM.Handler` (the transaction-level pipeline) and by system
   contracts that need to evaluate bytecode against an existing world state.
 
   ## EVM Concepts
